main
This application system will process data from 38 data from each district / city in East Java (Input Form) which will then be processed and compiled into an annual report (Output Template) in the form of a statistical book. However, the current application development is still a prototype for processing data on Budidaya Laut (Marine Aquaculture) only, which is expected to be developed in the future to manage all data.